+++ src/modules/palette/private/PaletteFormat.cpp
@@ -9,6 +9,7 @@
 #include "JASCPalette.h"
 #include "PNGPalette.h"
 #include "PaintNetPalette.h"
+#include "PhotoshopPalette.h"
 #include "QBCLPalette.h"
 #include "RGBPalette.h"
 #include "core/FourCC.h"
@@ -30,6 +31,8 @@ static core::SharedPtr<PaletteFormat> getFormat(const io::FormatDescription &des
 			return core::make_shared<ASEPalette>();
 		} else if (ext == "csv") {
 			return core::make_shared<CSVPalette>();
+		} else if (ext == "aco") {
+			return core::make_shared<PhotoshopPalette>();
 		} else if (ext == "txt") {
 			return core::make_shared<PaintNetPalette>();
 		} else if (ext == "png") {