+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ General:
 
    - Fixed invalid uv coordinate handling when importing meshes to voxelize them
    - Added support for paint.net palettes
+   - Added support for Adobe Photoshop ACO palettes
 
 VoxEdit: